# gamindie-ui/src/app/core/services/models/friend-ship-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FriendShipResponse {
  createdData?: string;
  id?: number;
  receiverAvatar?: string;
  receiverId?: number;
  receiverUsername?: string;
  senderAvatar?: string;
  senderId?: number;
  senderUsername?: string;
  status?: string;
}