# gamindie-ui/src/app/core/services/models/page-response-friend-ship-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FriendShipResponse } from '../models/friend-ship-response';
export interface PageResponseFriendShipResponse {
  content?: Array<FriendShipResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}